+++ CafeteriaApp/CafeteriaApp.Backend/Requests/OrderItem.php
@@ -3,41 +3,40 @@ require __DIR__ . '/../Controllers/OrderItem.php';
 require __DIR__ . '/TestRequestInput.php';
 
 if ($_SERVER['REQUEST_METHOD'] == 'GET') {
-    if (isset($_GET['orderId']) && testInt($_GET['orderId'])) // cashier only
-    {
-        checkResult(getOrderItemsByOrderId($conn, $_GET['orderId']));
+    if ( isset($_GET['orderId']) && testInt($_GET['orderId']) ) {// cashier only
+        checkResult( getOrderItemsByOrderId($conn, $_GET['orderId']) );
     }
 
 }
 
 if ($_SERVER['REQUEST_METHOD'] == 'POST') {
     //decode the json data
-    $data = json_decode(file_get_contents('php://input'));
+    $data = json_decode( file_get_contents('php://input') );
 
-    if (isset($data->OrderId, $data->MenuItemId, $data->Quantity) && testMutipleInts($data->OrderId, $data->MenuItemId, $data->Quantity)) {
+    if ( isset($data->OrderId, $data->MenuItemId, $data->Quantity) && testMutipleInts($data->OrderId, $data->MenuItemId, $data->Quantity) ) {
         $orderId = addOrderItem($conn, $data->OrderId, $data->MenuItemId, $data->Quantity);
 
-        if (!empty($orderId)) {
+        if ( !empty($orderId) ) {
             echo $orderId;
         }
-
-    } else {
+    }
+    else {
         echo 'error';
     }
 }
 
 if ($_SERVER['REQUEST_METHOD'] == 'PUT') {
     //decode the json data
-    $data = json_decode(file_get_contents('php://input'));
+    $data = json_decode( file_get_contents('php://input') );
 
-    if (isset($data->Id, $data->Quantity) && testMutipleInts($data->Id, $data->Quantity)) {
+    if ( isset($data->Id, $data->Quantity) && testMutipleInts($data->Id, $data->Quantity) ) {
         editOrderItemQuantity($conn, $data->Quantity, $data->Id);
     }
 
 }
 
-if ($_SERVER['REQUEST_METHOD'] == 'DELETE') {
-    if (isset($_GET['id']) && testInt($_GET['id'])) {
+if ( $_SERVER['REQUEST_METHOD'] == 'DELETE' && checkCSRFToken($_POST['csrf_token']) ) {
+    if ( isset($_GET['id']) && testInt($_GET['id']) ) {
         deleteOrderItem($conn, $_GET['id']);
     }
 }
